+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: lvargas- <lvargas-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 16:15:56 by lvargas-          #+#    #+#             */
-/*   Updated: 2025/04/08 17:39:26 by lvargas-         ###   ########.fr       */
+/*   Updated: 2025/04/09 15:40:18 by lvargas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, unsigned long n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned long	i;
 	unsigned char	*ptr;
